* Implement zboom

* Fix euro sign

* Remove phone length validation

* Fix zboom

* Change prices to BigDecimal

* Fix PriceAdapter.kt

* Fix home progress

* Fix localization

* Release 1.0.1 (9)

* Update dependencies

* Fix AuthCodeFragment sms retriever api permission

* Release 1.0.2 (10)

* Update dependencies

* Coupon popup (#40)

* Implement CouponReceivedDialogFragment.kt

* Fix code style

---------

+++ app/src/main/java/ru/frogogo/whitelabel/data/model/api/home/HomeProgress.kt
@@ -2,11 +2,12 @@ package ru.frogogo.whitelabel.data.model.api.home
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 
 @JsonClass(generateAdapter = true)
 data class HomeProgress(
   @Json(name = "current")
-  val current: Int,
+  val current: BigDecimal,
   @Json(name = "target")
-  val target: Int,
+  val target: BigDecimal,
 )
