+++ app/src/main/java/ru/frogogo/whitelabel/data/model/api/home/HomePromotion.kt
@@ -3,6 +3,7 @@ package ru.frogogo.whitelabel.data.model.api.home
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import ru.frogogo.whitelabel.data.model.api.Photo
+import java.math.BigDecimal
 
 @JsonClass(generateAdapter = true)
 data class HomePromotion(
@@ -13,7 +14,7 @@ data class HomePromotion(
   @Json(name = "steps")
   val steps: List<String>,
   @Json(name = "price")
-  val price: Int,
+  val price: BigDecimal,
   @Json(name = "discounted_price")
-  val discountedPrice: Int,
+  val discountedPrice: BigDecimal,
 )
